+++ cx-portal/src/features/apps/mapper.ts
@@ -57,3 +57,19 @@ export const filterSubscribed = (
     .filter((app: AppMarketplaceApp) => subscribed?.includes(app.id))
     .map(appToCard)
 }
+
+export const appToStatus = (
+  apps: AppMarketplaceApp[],
+  subscriptionStatus: SubscriptionStatusItem[]
+): AppMarketplaceApp[] => {
+  return apps?.map((app: AppMarketplaceApp) => {
+    const status =
+      subscriptionStatus?.find((e) => e.appId === app.id)
+        ?.appSubscriptionStatus || SubscriptionStatus.INACTIVE
+    const image = {
+      src: getAppLeadImage(app),
+      alt: app.title,
+    }
+    return { ...app, status, image }
+  })
+}